refactor: move ExtensionConfig to shared core package
- Move ExtensionConfig interface from duplicated locations to packages/core/src/config/config.ts
- Update CLI package to import ExtensionConfig from @google/gemini-cli-core
- Update A2A-server package to import ExtensionConfig from @google/gemini-cli-core
- Fix Extension interface in a2a-server to include missing path property
- Eliminate code duplication and ensure consistent interface across packages
- Addresses code review feedback about ExtensionConfig duplication

This resolves the architectural issue where ExtensionConfig was duplicated
between CLI and A2A-server packages, causing inconsistencies like missing
excludeTools property in the a2a-server version.
